requirements.txt
alembic==1.13.2
amulet-core==1.9.25
amulet-leveldb==1.0.2
amulet-map-editor==0.10.33
amulet-nbt==2.1.3
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-mailman==1.1.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
lz4==4.3.3
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
minecraft-resource-pack==1.4.3
mutf8==1.0.6
numpy==1.26.4
packaging==24.1
pillow==10.4.0
platformdirs==3.11.0
portalocker==2.10.0
psycopg2-binary==2.9.9
PyMCTranslate==1.2.27
PyOpenGL==3.1.7
python-dotenv==1.0.1
python-http-client==3.3.7
pythoncms==1.3.3
pywatchman==2.0.0
requests==2.31.0
sendgrid==6.11.0
shopyo==4.8.6
six==1.16.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.41.2
starkbank-ecdsa==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel @ file:///home/linuxbrew/.linuxbrew/Cellar/python%403.12/3.12.3/libexec/wheel-0.43.0-py3-none-any.whl#sha256=80a49545e5e5ef2abeb79ec254447300d41dfb76914fc65efa6da8dd0a766052
WTForms==3.1.2
wxPython==4.2.1
